Add basic processes and threads

+++ echis/include/echis/mm/vm/node.h
@@ -9,7 +9,7 @@ namespace echis
     {
         struct VMAllocNode
         {
-            std::uint32_t numPages;
+            std::uint64_t numPages;
             std::uint64_t base;
         };
     }
